Implement generate random world method in the main controller, and fix issue where the templayers need to be floats not ints

+++ DynamicCavemanWorld/Assets/Classes/Editor/GeneratorTester.cs
@@ -42,7 +42,7 @@ public class GeneratorTester {
         // Set the world size
         DataGenerator generator = new DataGenerator(40, 60);
         // Generate the tested world
-        int[][,] array = generator.CreateTemperatureLayers(4);
+        float[][,] array = generator.CreateTemperatureLayers(4);
         // Generate random coordinates
         System.Random randy = new System.Random();
         int i = randy.Next(0, generator.WORLDX);
